Fix float64 warning on MPS by casting alphas to input dtype
np.logspace returns float64 by default, which triggers the torch_mps
float64-to-float32 cast warning even when input data is already float32.

+++ benchmarks/benchmark_backends.py
@@ -281,7 +281,7 @@ def run_benchmarks(args):
     X_train, X_test = X[:split], X[split:]
     Y_train, Y_test = Y[:split], Y[split:]
 
-    alphas = np.logspace(-3, 3, args.n_alphas).tolist()
+    alphas = np.logspace(-3, 3, args.n_alphas, dtype=dtype).tolist()
 
     backends = args.backends
     print(f"Backends to benchmark: {backends}")
